+++ tests/acceptance/bi-service-sdk-as-plugin.js
@@ -1,44 +1,29 @@
-const packageJSON  = require('../../../package.json');
 const Promise      = require('bluebird');
 const fs           = Promise.promisifyAll(require('fs'));
 const path         = require('path');
-const sinon        = require('sinon');
 const chai         = require('chai');
-const sinonChai    = require("sinon-chai");
 const childProcess = require('child_process');
-const tmp          = require('tmp');
 
-var expect = chai.expect;
+const expect = chai.expect;
 
-chai.use(sinonChai);
 chai.should();
 
-var BI_SERVICE_DOC_EXEC = require.resolve('bi-service-doc/bin/bi-service-doc');
-
-describe('bin/bi-service-sdk', function() {
-    before(function(done) {
-        var self = this;
-
-        tmp.setGracefulCleanup();
-        this.tmpDir = tmp.dirSync({
-            //unsafeCleanup: true,
-            keep: true
-        });
+describe('bi-service-sdk as a plugin', function() {
+    before(function() {
+        const self = this;
 
         this.spawn = spawn;
 
-
-        this.sdk1FilePath = `${this.tmpDir.name}/bi-service-app/bi-test-app1-1.0.0.zip`;
-        this.sdk2FilePath = `${this.tmpDir.name}/bi-service-app/bi-test-app2-1.0.0.zip`;
-
-        prepareProjectTestCase(this.tmpDir.name + '/bi-service-app', done);
+        this.rootPath = path.resolve(__dirname + '/../bi-service-app-2');
+        this.sdk1FilePath = path.resolve(this.rootPath + `/bi-test-app1-1.0.0.zip`);
+        this.sdk2FilePath = path.resolve(this.rootPath + `/bi-test-app2-1.0.0.zip`);
 
         function spawn(args) {
-            var cmd = path.normalize(__dirname + '/../../../bin/bi-service-sdk.js');
+            const cmd = path.normalize(__dirname + '/../../node_modules/.bin/bi-service');
             args.unshift(cmd);
 
-            var result = childProcess.spawnSync('node', args, {
-                cwd: self.tmpDir.name + '/bi-service-app',
+            const result = childProcess.spawnSync('node', args, {
+                cwd: self.rootPath,
                 env: { NODE_ENV: 'development' }
             });
 
@@ -48,58 +33,14 @@ describe('bin/bi-service-sdk', function() {
 
             return result;
         }
-
-        function prepareProjectTestCase(dest, cb) {
-            let appPath = path.resolve(__dirname + '/../../bi-service-app');
-
-            if (!fs.existsSync(dest)) {
-                fs.mkdirSync(dest);
-            }
-
-            return Promise.all([
-                fs.readFileAsync(`${appPath}/config.json5`),
-                fs.readFileAsync(`${appPath}/index.js`),
-                fs.readFileAsync(`${appPath}/package.json`)
-            ]).then(function(files) {
-                return Promise.all([
-                    fs.writeFileAsync(`${dest}/config.json5`, files[0]),
-                    fs.writeFileAsync(`${dest}/index.js`, files[1]),
-                    fs.writeFileAsync(`${dest}/package.json`, files[2])
-                ]);
-            }).then(function() {
-                const result = childProcess.spawnSync('npm', ['install'], {
-                    cwd: dest,
-                    env: { NODE_ENV: 'development' }
-                });
-
-                if (result.error) {
-                    throw result.error;
-                } else if (result.status !== 0) {
-                    throw new Error('npm install err: ' + result.stderr.toString());
-                }
-            }).asCallback(cb);
-        }
-    });
-
-    it('should dump bi-service-sdk version to stdout and exit with code: 0', function() {
-        var result = this.spawn(['--version']);
-
-        result.status.should.be.equal(0);
-        result.stdout.toString().should.match(new RegExp(packageJSON.version));
     });
 
     describe('', function() {
         before(function() {
             this.result = this.spawn([
-                '--service',
-                path.resolve(this.tmpDir.name + '/bi-service-app'),
-                '--doc-exec',
-                './node_modules/.bin/bi-service-doc',
-                '--tests',
-                true,
-                '--',
+                'build:sdk',
                 '--config',
-                path.resolve(this.tmpDir.name + '/bi-service-app/config.json5')
+                path.resolve(this.rootPath + '/config.json5')
             ]);
         });
 
@@ -141,17 +82,13 @@ describe('bin/bi-service-sdk', function() {
     describe('--dry option', function() {
         it('should not create any SDK packages if dry run is enabled', function() {
             var result = this.spawn([
-                '--service',
-                path.resolve(this.tmpDir.name + '/bi-service-app'),
-                '--doc-exec',
-                require.resolve('bi-service-doc/bin/bi-service-doc'),
+                'build:sdk',
                 '--tests',
                 false,
                 '--dry',
                 true,
-                '--',
                 '--config',
-                path.resolve(this.tmpDir.name + '/bi-service-app/config.json5')
+                path.resolve(this.rootPath + '/config.json5')
             ]);
 
             result.status.should.be.equal(0);
@@ -164,19 +101,15 @@ describe('bin/bi-service-sdk', function() {
     describe('--cleanup option', function() {
         it('should NOT remove tmp build directory when the `cleanup` option is set to false', function() {
             var result = this.spawn([
-                '--service',
-                path.resolve(this.tmpDir.name + '/bi-service-app'),
-                '--doc-exec',
-                require.resolve('bi-service-doc/bin/bi-service-doc'),
+                'build:sdk',
                 '--tests',
                 false,
                 '--dry',
                 true,
                 '--cleanup',
                 false,
-                '--',
                 '--config',
-                path.resolve(this.tmpDir.name + '/bi-service-app/config.json5')
+                path.resolve(this.rootPath + '/config.json5')
             ]);
 
             result.status.should.be.equal(0);