+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.0
+
+* [ADDED] - `--specs` option to `bi-service-sdk` executable
+* [ADDED] - make use of bi-service shell API - `build:sdk` command is available via `bi-service` executable
+
 ## 1.1.3
 
 * [FIXED] - broken prototype inheritance